demos/6-observability-signals/app/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-instrumentation-logging==0.46b0
opentelemetry-instrumentation-requests==0.46b0
opentelemetry-exporter-otlp==1.25.0
